Lecture 04 examples

+++ 03/String.h
@@ -13,6 +13,6 @@ namespace samples
 
 	private:
 		char* mString;
-		int mSize;
+		size_t mSize;
 	};
 }
